+++ src/main/java/nextstep/AppConfig.java
@@ -1,0 +1,10 @@
+package nextstep;
+
+import auth.AuthConfig;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.context.annotation.Import;
+
+@Configuration
+@Import(value = {AuthConfig.class})
+public class AppConfig {
+}
